refactor(crux-ui): move deployment related components to the root deployments folder
