+++ web/crux-ui/src/components/deployments/copy-deployment-card.tsx
