+++ web/crux-ui/src/components/deployments/deployment-details-card.tsx
