+++ web/crux-ui/src/components/deployments/deployment-details-section.tsx
