+++ web/crux-ui/src/components/deployments/deployment-view-list.tsx
