+++ web/crux-ui/src/components/deployments/deployment-view-tile.tsx
