+++ web/crux-ui/src/components/projects/versions/version-sections.tsx
@@ -4,8 +4,8 @@ import { ProjectDetails, VERSION_SECTIONS_STATE_VALUES } from '@app/models'
 import { parseStringUnionType } from '@app/utils'
 import { useRouter } from 'next/dist/client/router'
 import React, { useEffect, useRef } from 'react'
-import AddDeploymentToVersionCard from './deployments/add-deployment-to-version-card'
-import CopyDeploymentCard from './deployments/copy-deployment-card'
+import AddDeploymentToVersionCard from '../../deployments/add-deployment-to-version-card'
+import CopyDeploymentCard from '../../deployments/copy-deployment-card'
 import AddImagesCard from './images/add-images-card'
 import { VerionState, VersionActions, VersionSection } from './use-version-state'
 import VersionDeploymentsSection from './version-deployments-section'
