+++ web/crux-ui/src/pages/[teamSlug]/deployments.tsx
@@ -1,10 +1,8 @@
 import AddDeploymentCard from '@app/components/deployments/add-deployment-card'
+import CopyDeploymentCard from '@app/components/deployments/copy-deployment-card'
+import DeploymentStatusTag, { deploymentStatusTranslation } from '@app/components/deployments/deployment-status-tag'
+import useCopyDeploymentState from '@app/components/deployments/use-copy-deployment-state'
 import { Layout } from '@app/components/layout'
-import CopyDeploymentCard from '@app/components/projects/versions/deployments/copy-deployment-card'
-import DeploymentStatusTag, {
-  deploymentStatusTranslation,
-} from '@app/components/projects/versions/deployments/deployment-status-tag'
-import useCopyDeploymentState from '@app/components/projects/versions/deployments/use-copy-deployment-state'
 import { BreadcrumbLink } from '@app/components/shared/breadcrumb'
 import Filters from '@app/components/shared/filters'
 import PageHeading from '@app/components/shared/page-heading'
