+++ web/crux-ui/src/pages/[teamSlug]/deployments/[deploymentId].tsx
@@ -1,13 +1,13 @@
 import EditorBadge from '@app/components/editor/editor-badge'
 import { Layout } from '@app/components/layout'
 import NodeConnectionCard from '@app/components/nodes/node-connection-card'
-import CopyDeploymentCard from '@app/components/projects/versions/deployments/copy-deployment-card'
-import CreateDeploymentTokenCard from '@app/components/projects/versions/deployments/create-deployment-token-card'
-import DeploymentDetailsSection from '@app/components/projects/versions/deployments/deployment-details-section'
-import DeploymentTokenCard from '@app/components/projects/versions/deployments/deployment-token-card'
-import EditDeploymentCard from '@app/components/projects/versions/deployments/edit-deployment-card'
-import EditDeploymentInstances from '@app/components/projects/versions/deployments/edit-deployment-instances'
-import useDeploymentState from '@app/components/projects/versions/deployments/use-deployment-state'
+import CopyDeploymentCard from '@app/components/deployments/copy-deployment-card'
+import CreateDeploymentTokenCard from '@app/components/deployments/create-deployment-token-card'
+import DeploymentDetailsSection from '@app/components/deployments/deployment-details-section'
+import DeploymentTokenCard from '@app/components/deployments/deployment-token-card'
+import EditDeploymentCard from '@app/components/deployments/edit-deployment-card'
+import EditDeploymentInstances from '@app/components/deployments/edit-deployment-instances'
+import useDeploymentState from '@app/components/deployments/use-deployment-state'
 import { BreadcrumbLink } from '@app/components/shared/breadcrumb'
 import PageHeading from '@app/components/shared/page-heading'
 import { DetailsPageMenu } from '@app/components/shared/page-menu'
