+++ web/crux-ui/src/pages/[teamSlug]/deployments/[deploymentId]/instances/[instanceId].tsx
@@ -2,8 +2,8 @@ import EditorBadge from '@app/components/editor/editor-badge'
 import useEditorState from '@app/components/editor/use-editor-state'
 import useItemEditorState from '@app/components/editor/use-item-editor-state'
 import { Layout } from '@app/components/layout'
-import useInstanceState from '@app/components/projects/versions/deployments/instances/use-instance-state'
-import useDeploymentState from '@app/components/projects/versions/deployments/use-deployment-state'
+import useInstanceState from '@app/components/deployments/instances/use-instance-state'
+import useDeploymentState from '@app/components/deployments/use-deployment-state'
 import CommonConfigSection from '@app/components/projects/versions/images/config/common-config-section'
 import configToFilters from '@app/components/projects/versions/images/config/config-to-filters'
 import CraneConfigSection from '@app/components/projects/versions/images/config/crane-config-section'
